Add flight auth integration test
