+++ crates/runtime/Cargo.toml
@@ -135,6 +135,7 @@ azure_storage_blobs = "0.20.0"
 azure_core = "0.20.0"
 azure_storage = "0.20.0"
 scopeguard = "1.2.0"
+flightrepl = { path = "../flightrepl" }
 
 [features]
 aws-secrets-manager = [
